Updated test setup
