+++ mmeds/html/welcome.html
@@ -2,18 +2,18 @@
 <html>
   <body>
     <div class='inner'>
-      <h3>Access Previous Uploads</h3>
       <p>
+      <h3>Access Previous Uploads</h3>
         <form action='download_page' method='post'>
           <p>
             <input type="text" placeholder="Access Code" name="access_code" required/>
-          </p>
           <button type="submit">Execute</button>
+          </p>
         </form>
       </p>
-      <h3>Convert Metadata File</h3>
       <p>
-        <form action='convert_metadata' id='convert_form' method='post' enctype='multipart/form-data'>
+      <h3>Convert Metadata File</h3>
+        <form action='convert_metadata' id='convert_form' formmethod='post' enctype='multipart/form-data'>
           <p>
             <h4> Convert file to </h4>
             <select name='convertTo' placeholder='Convert file to' form='convert_form'>
@@ -23,18 +23,18 @@
             <input type='file' placeholder='MIxS metadata file' name='myMetaData' required/> <br />
             <input type='text' placeholder='Unit column header' name='unitCol'> <br />
             <input type='number' placeholder='Number of rows to skip' name='skipRows'> <br />
+          <button type='submit' formmethod='post'>Convert File</button>
           </p>
-          <button type='submit' method='post'>Convert File</button>
         </form>
       </p>
-      <h3>Forgot Access Code</h3>
       <p>
-        <form action='reset_code' method='post'>
+      <h3>Forgot Access Code</h3>
+        <form action='reset_code' formmethod='post'>
           <p>
             <input type="text" placeholder="Study Name" name="study_name" required/>
             <input type="email" placeholder="Email" name="study_email" required/>
+            <button type="submit">Execute</button>
           </p>
-          <button type="submit">Execute</button>
         </form>
       </p>
     </div>
