+++ mmeds/tests/unit/test_database.py
@@ -155,7 +155,7 @@ class DatabaseTests(TestCase):
             # Otherwise check the absolute value of the difference is small
             # so that SQL won't fail to match floats
             else:
-                sql += ' ABS(' + table + '.' + column + ' - ' + str(value) + ') <= 0.01'
+                sql += '{} = {}'.format(column, value)
         if table in fig.PROTECTED_TABLES:
             sql += ' AND user_id = ' + str(self.user_id)
 
@@ -218,40 +218,6 @@ class DatabaseTests(TestCase):
                 jresult = self.c.execute(sql)
                 # Ensure an entry exists for this value
                 assert jresult > 0
-
-    def error_test_modify_tables(self):
-        self.c.execute('SHOW TABLES')
-        tables = [x[0] for x in self.c.fetchall() if 'protected' not in x[0]]
-        del tables[tables.index('session')]
-        del tables[tables.index('security_token')]
-        del tables[tables.index('user')]
-        for table in tables:
-            sql = 'DESCRIBE {}'.format(table)
-            self.c.execute(sql)
-            result = self.c.fetchall()
-            columns = [res[0] for res in result]
-
-            sql = 'SELECT * FROM {}'.format(table)
-            self.c.execute(sql)
-            rows = self.c.fetchall()
-            # Pick a row at random
-            row = random.choice(rows)
-
-            sql = 'DELETE FROM {} WHERE '.format(table)
-            for i, column in enumerate(columns):
-                value = row[i]
-                if i == 0:
-                    sql += ' '
-                else:
-                    sql += ' AND '
-                # Add qoutes around string values
-                if type(value) == str:
-                    sql += column + ' = "' + value + '"'
-                # Otherwise check the absolute value of the difference is small
-                # so that SQL won't fail to match floats
-                else:
-                    sql += ' ABS(' + table + '.' + column + ' - ' + str(value) + ') <= 0.01'
-            self.c.execute(sql)
 
     def test_c_table_protection(self):
         """
